requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.42.0=pypi_0
app=0.1=dev_0
boto3=1.35.97=pypi_0
botocore=1.35.97=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.12.31=h06a4308_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
deprecation=2.1.0=pypi_0
distro=1.9.0=pypi_0
eval-type-backport=0.2.2=pypi_0
google-auth=2.37.0=pypi_0
gotrue=2.11.1=pypi_0
griffe=1.5.4=pypi_0
groq=0.15.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.27.2=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.10=pypi_0
jiter=0.8.2=pypi_0
jmespath=1.0.1=pypi_0
jsonpath-python=1.0.6=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
logfire-api=3.1.0=pypi_0
mistralai=1.2.6=pypi_0
ncurses=6.4=h6a678d5_0
openai=1.59.6=pypi_0
openssl=3.0.15=h5eee18b_0
pip=24.2=py310h06a4308_0
postgrest=0.19.1=pypi_0
propcache=0.2.1=pypi_0
pydantic=2.10.4=pypi_0
pydantic-ai=0.0.18=pypi_0
pydantic-ai-slim=0.0.18=pypi_0
pydantic-core=2.27.2=pypi_0
pytest=8.3.4=pypi_0
python=3.10.16=he870216_1
python-dotenv=1.0.1=pypi_0
readline=8.2=h5eee18b_0
realtime=2.1.0=pypi_0
replicate=1.0.4=pypi_0
requests=2.32.3=pypi_0
s3transfer=0.10.4=pypi_0
setuptools=75.1.0=py310h06a4308_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
storage3=0.11.0=pypi_0
strenum=0.4.15=pypi_0
stripe=11.4.1=pypi_0
supabase=2.11.0=pypi_0
supafunc=0.9.0=pypi_0
tk=8.6.14=h39e8969_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
urllib3=2.3.0=pypi_0
wheel=0.44.0=py310h06a4308_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
zlib=1.2.13=h5eee18b_1